REORGANIZE: Rename folder
I renamed a specific folder from subgraph_matching to cbir_subsg.
Therefore, some codes also was modified.
