+++ cbir_subsg/config.py
